requirements.txt
distlib==0.3.0
Django==3.0.5
django-crispy-forms==1.9.0
docutils==0.16
filelock==3.0.12
gitdb==4.0.4
GitPython==3.1.1
idna==2.9
importlib-metadata==1.6.0
jedi==0.17.0

pathtools==0.1.2
pexpect==4.8.0
Pillow==7.1.2
psycopg2==2.8.5
ptyprocess==0.6.0
pycryptodome==3.9.7

pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
requests==2.23.0
sh==1.12.14
six==1.14.0
smmap==3.0.2
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.18
django-taggit==1.3.0
django-nested-admin==3.3.2